Crate odds

Odds and ends — collection miscellania.

The goal of this crate is to abolish itself. Things that are here will move to other places when possible.

The odds crate has the following cargo feature flags:

unstable.

Optional.

Requires nightly channel.

Implement the closure traits for Fix.

Structs

Fix
Fixpoint combinator for rust closures, generalized over the return type.

Traits

IndexRange
IndexRange is implemented by Rust's built-in range types, produced by range syntax like .., a.., ..b or c..d.

Functions

debug_assert_unreachable
Act as debug_assert! in debug mode, asserting that this point is not reached.

get_unchecked
Use debug_assert! to check indexing in debug mode. In release mode, no checks are done.

get_unchecked_mut
Use debug_assert! to check indexing in debug mode. In release mode, no checks are done.

ptr_eq
Compare if a and b are equal pointers.

raw_byte_repr
Safe to use with any wholly initialized memory ptr

ref_eq
Compare if a and b are equal as pointers.

slice_unchecked
Check slicing bounds in debug mode, otherwise just act as an unchecked slice call.