Function odds::debug_assert_unreachable

pub unsafe fn debug_assert_unreachable() -> !

Act as debug_assert! in debug mode, asserting that this point is not reached.

In release mode, no checks are done, and it acts like the unreachable intrinsic.