Module odds::slice

Extra functions for slices

Modules

iter

unalign

Structs

MendSlices
An iterator adaptor that glues together adjacent contiguous slices.

RevSlice
A reversed view of a slice.

Traits

MendSlice
A trait for items that can maybe be joined together.

Pod
"plain old data": Types that we can stick arbitrary bit patterns into, and thus use them as blocks in split_aligned_for or in UnalignedIter.

SliceFind
Element-finding methods for slices

SliceFindSplit
Element-finding methods for slices

SliceIterExt
Extra iterator adaptors for iterators of slice elements.

Functions

rotate_left
Rotate steps towards lower indices.

shared_prefix
Return the end index of the longest shared (equal) prefix of a and b.

split_aligned_for
Split the input slice into three chunks, so that the middle chunk is a slice of a larger "block size" (for example T could be u64) that is correctly aligned for T.