Trait odds::string::StrSlice

pub trait StrSlice {
    fn get_slice<R>(&self, r: R) -> Option<&str> where R: IndexRange;
}

Extension trait for str for string slicing without panicking

Required Methods

fn get_slice<R>(&self, r: R) -> Option<&str> where R: IndexRange

Return a slice of the string, if it is in bounds /and on character boundaries/, otherwise return None