Trait odds::vec::VecExt

pub trait VecExt<T> {
    fn splice<R, I>(&mut self, r: R, iter: I) where I: IntoIterator<Item=T>, I::IntoIter: ExactSizeIterator, R: IndexRange;
}

Extra methods for Vec<T>

Required Methods

fn splice<R, I>(&mut self, r: R, iter: I) where I: IntoIterator<Item=T>, I::IntoIter: ExactSizeIterator, R: IndexRange

Remove elements in a range, and insert from an iterator in their place.

The removed and inserted ranges don't have to match in length.

Panics if range r is out of bounds.

Panics if iterator iter is not of exact length.