Struct odds::Fix

pub struct Fix<'a, T: 'a, R: 'a = T>(pub &'a Fn(Fix<T, R>, T) -> R);

Fixpoint combinator for rust closures, generalized over the return type.

In Fix<T, R>, T is the argument type, and R is the return type, R defaults to T.

Calling the Fix value only supports function call notation with the nightly channel and the crate feature ‘unstable’ enabled; use the .call() method otherwise.

Use this best through the fix function.

use odds::Fix;
use odds::fix;

let c = |f: Fix<i32>, x| if x == 0 { 1 } else { x * f.call(x - 1) };
let fact = Fix(&c);
assert_eq!(fact.call(5), 120);

let data = [true, false];
assert!(!fix(&data[..], |f, x| {
    x.len() == 0 || x[0] && f.call(&x[1..])
}));

Methods

impl<'a, T, R> Fix<'a, T, R>

fn call(&self, arg: T) -> R